Fixed issues and moved i18n to separate file.
