+++ js/masters/modules/accessibility.src.js
@@ -8,5 +8,6 @@
  * License: www.highcharts.com/license
  */
 'use strict';
+import '../../modules/a11y-i18n.src.js';
 import '../../modules/screen-reader.src.js';
 import '../../modules/keyboard-navigation.src.js';
